+++ src/descriptor.rs
@@ -1,7 +1,7 @@
-mod data_package;
+mod dataset;
 mod table_schema;
 
-pub use data_package::{Dataset, Name, Table, TableSource};
+pub use dataset::{Dataset, Name, Table, TableSource};
 pub use table_schema::{
     AnyFieldType, ArrayFieldType, BooleanFieldType, Constraints, DateFieldType, DateTimeFieldType,
     GeoPointFieldType, NumberFieldType, ObjectFieldType, StringFieldFormat, StringFieldType,
